backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
psycopg[binary]>=3.1
pydantic>=2.5
python-dotenv>=1.0
requests>=2.31